Go-SDK: exclude oversized-payload framing test from -race builds
TestWriteFrameRejectsOversizedPayload uses unsafe.Slice with a
length greater than MaxFrameSize on a single-byte backing array to
exercise writeFrame's size guard without actually allocating 4 GiB.
Under -race, Go's checkptr instrumentation treats the resulting
slice as straddling allocations and fatals the entire test binary,
which currently prevents go test -race from running on the
pkg/execution package at all.

Move the test into its own file gated by //go:build !race so the
non-race build keeps the coverage and the race build skips it. The
guard under test is a single len() comparison; losing race-mode
coverage on that one path is acceptable to keep the rest of the
package's race tests runnable.
